Move txs rpc into sequencer
