+++ examples/demo-stf/src/batch_builder.rs
@@ -89,6 +89,7 @@ where
 
             // Execute
             {
+                // TODO: Bug(!), because potential discrepancy with default stf or any other STF
                 let sender_address: C::Address = tx.pub_key().to_address();
                 let ctx = C::new(sender_address);
