+++ rollup-interface/src/node/services/stf_runner.rs
@@ -20,6 +20,7 @@ use crate::zk::traits::Zkvm;
 ///
 /// and a `impl StateTransitionRunner<ZkConfig, Vm> for MyRunner` which instead uses a state root as its runtime config.
 ///
+/// TODO: Why is it called runner? It only creates. Creator, Factory.
 pub trait StateTransitionRunner<T: StateTransitionConfig, Vm: Zkvm> {
     /// The parameters of the state transition function which are set at runtime. For example,
     /// the runtime config might contain path to a data directory.
